refactor: update base urls with env
