+++ .env.example
@@ -1,0 +1,8 @@
+URL_GENRE=http://localhost:3001/v1/genres
+URL_ARTIST=http://localhost:3002/v1/artists
+URL_BAND=http://localhost:3003/v1/bands
+URL_USER=http://localhost:3004/v1/users
+URL_ALBUM=http://localhost:3005/v1/albums
+URL_TRACK=http://localhost:3006/v1/tracks
+URL_FAVOURITES=http://localhost:3007/v1/favourites
+PORT=4000
